fix: lint and avoid monkey patch dunder new
